backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Samuel Miranda <samuelmirandasamt@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13"
django = ">=5.2.1,<6.0.0"
djangorestframework = ">=3.16.0,<4.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
dj-database-url = ">=2.3.0,<3.0.0"
drf-yasg = ">=1.21.10,<2.0.0"
pytest = ">=8.3.5,<9.0.0"
pytest-django = ">=4.11.1,<5.0.0"
black = ">=25.1.0,<26.0.0"
isort = ">=6.0.1,<7.0.0"
flake8 = ">=7.2.0,<8.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"